+++ domain/src/main/java/com/sangmeebee/searchmovie/domain/model/Movie.kt
@@ -3,7 +3,7 @@ package com.sangmeebee.searchmovie.domain.model
 data class Movie(
     val offsetTime: String,
     val totalCount: Long,
-    val pageStart: Long,
+    val pageStart: Int,
     val pageSize: Int,
     val items: List<MovieInfo>,
 )
